listen live, smash team, social media components added
