+++ src/app/access/access.component.html
@@ -1,4 +1,4 @@
-<div class="jumbotron">
+<div class="jumbotron jumbotron-fluid">
   <div class="container">
     <div class="card">
       <div class="card-block">
