+++ src/app/access/access.component.ts
@@ -8,7 +8,15 @@ import { Component, OnInit } from '@angular/core';
       background-image: url('../assets/background-1.jpg');
       background-size: cover;
       background-repeat: no-repeat;
+      background-position: right;
+      background-attachment: fixed;
 
+      margin-bottom: 0;
+    }
+
+    .card {
+      background-color: rgba(255, 255, 255, 0.8);
+      height: 200px;
     }
 
   `]
