--- src/app/home/contact/contact.component.html
@@ -1,3 +1,0 @@
-<p>
-  contact works!
-</p>
