+++ src/app/home/home.component.css
@@ -3,7 +3,5 @@
   min-height: 300px;
   font-style: italic;
 }
-.hSkyscrapperBg {
-  background-image: url("../../assets/skyscrappers/skyscrapperBackground.jpg");
-  background-size: contain;
-}
+
+
