+++ src/app/home/home.component.html
@@ -1,6 +1,9 @@
 <app-access></app-access>
+<app-listen-live></app-listen-live>
+<app-smash-team></app-smash-team>
+<app-social-media></app-social-media>
 
-<div class="container">
+<!--<div class="container">
   <div *ngFor="let post of posts" class="card">
     <h3 class="card-header">{{post.id}}</h3>
     <div class="card-block">
@@ -9,142 +12,4 @@
     </div>
 
   </div>
-</div>
-
-
-
-
-<!--<div class="jumbotron jumbotron-fluid">
-  <div class="container">
-    <h1 class="display-3">Fluid jumbotron</h1>
-    <p class="lead">This is a modified jumbotron that occupies the entire horizontal space of its parent.</p>
-  </div>
 </div>-->
-
-<!--<div id="carouselExampleIndicators" class="carousel slide" data-ride="carousel">
-  <ol class="carousel-indicators">
-    <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active"></li>
-    <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
-    <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
-  </ol>
-  <div class="carousel-inner" role="listbox">
-    <div class="carousel-item active">
-      <img class="d-block" src="{{jumbotronBanners.banner}}" alt="First slide">
-    </div>
-    <div class="carousel-item">
-      <img class="d-block img-fluid" src="{{jumbotronBanners.banner}}" alt="Second slide">
-    </div>
-    <div class="carousel-item">
-      <img class="d-block img-fluid" src="{{jumbotronBanners.banner}}" alt="Third slide">
-    </div>
-  </div>
-  <a class="carousel-control-prev" href="#carouselExampleIndicators" role="button" data-slide="prev">
-    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-    <span class="sr-only">Previous</span>
-  </a>
-  <a class="carousel-control-next" href="#carouselExampleIndicators" role="button" data-slide="next">
-    <span class="carousel-control-next-icon" aria-hidden="true"></span>
-    <span class="sr-only">Next</span>
-  </a>
-</div>-->
-
-<!--<carousel>
-  <slide>
-    <img src="assets/images/nature/1.jpg" alt="First slide">
-  </slide>
-  <slide>
-    <img src="assets/images/nature/2.jpg" alt="Second slide">
-  </slide>
-  <slide>
-    <img src="assets/images/nature/3.jpg" alt="Third slide">
-  </slide>
-</carousel>-->
-
-
-<!--<carousel [interval]=3000 [noWrap]="noWrapSlides">
-  <slide *ngFor="let jumbotronBanner of JumbotronBanners;let index=index">
-    <img [src]="jumbotronBanner.banner">
-    <div class="carousel-caption">
-      <h4>Slide {{index}}</h4>
-      <p>{{jumbotronBanner.subtitle}}</p>
-    </div>
-  </slide>
-  <i class="fa fa-chevron-left left carousel-control"></i>
-  <i class="fa fa-chevron-right right carousel-control"></i>
-</carousel>-->
-
-
-<!--<hr>
-<div class="container-fluid">
-  <div class="row align-items-start container-fluid px-0">
-    <div class="col-md-8">
-      <app-adblock-pnl class=""></app-adblock-pnl>
-      <div class="row">
-        <app-half-portrait-pnl class="col-6 mb-2"></app-half-portrait-pnl>
-        <app-half-portrait-pnl class="col-6 mb-2"></app-half-portrait-pnl>
-        <app-half-portrait-pnl class="col-6 mb-2"></app-half-portrait-pnl>
-        <app-half-portrait-pnl class="col-6 mb-2"></app-half-portrait-pnl>
-      </div>
-    </div>
-    <app-skyscrapper-pnl class=""></app-skyscrapper-pnl>
-  </div>
-
-
-
-  <div class="row align-items-start">
-    <app-banner-pnl class="col-md-12"></app-banner-pnl>
-
-    <div id="dRow-1" class="row col-md-12">
-      <div id="clmn-1" class="col-md-6">
-        <app-portrait-pnl class=""></app-portrait-pnl>
-      </div>
-
-      <div id="clmn-2" class="col-md-6">
-        <div class="row col-md-12">
-          <div class="col-md-6">
-            <div class="row">
-              <app-call2action-pnl class="col-md-6 d-inline-block"></app-call2action-pnl>
-              <app-call2action-pnl class="col-md-6 d-inline-block"></app-call2action-pnl>
-            </div>
-            <app-half-portrait-pnl class="col-md-12 d-inline"></app-half-portrait-pnl>
-            &lt;!&ndash;<div class="row">
-              <app-call2action-pnl class="col-md-6 d-inline-block"></app-call2action-pnl>
-              <app-call2action-pnl class="col-md-6 d-inline-block"></app-call2action-pnl>
-            </div>&ndash;&gt;
-          </div>
-          <div class="col-md-6">
-            <app-portrait-pnl class=""></app-portrait-pnl>
-          </div>
-        </div>
-        <div class="row col-md-12">
-          <app-adblock-pnl class=""></app-adblock-pnl>
-        </div>
-      </div>
-
-    </div>
-    <app-adblock-pnl class="col"></app-adblock-pnl>
-  </div>
-</div>-->
-<!--<div class="container-fluid">
-  <div class="row">
-    <div class="col-sm-6">
-      <div class="card">
-        <div class="card-block">
-          <h3 class="card-title">Special title treatment</h3>
-          <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
-          <a href="#" class="btn btn-primary">Go somewhere</a>
-        </div>
-      </div>
-    </div>
-    <div class="col-sm-6">
-      <div class="card">
-        <div class="card-block">
-          <h3 class="card-title">Special title treatment</h3>
-          <p class="card-text">With supporting text below as a natural lead-in to additional content.</p>
-          <a href="#" class="btn btn-primary">Go somewhere</a>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>-->
-
