+++ src/app/home/home.module.ts
@@ -2,10 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
+import { MODULE_ROUTES_COMPONENTS, MODULE_ROUTES } from '../app-routing.module';
 import { CarouselModule } from 'ng2-bootstrap/carousel/carousel.module';
 import { AccessModule } from '../access/access.module';
-import { MODULE_ROUTES_COMPONENTS, MODULE_ROUTES } from '../app-routing.module';
 
+import { ListenLiveComponent } from '../components/listen-live.component';
+import { SmashTeamComponent } from '../components/smash-team.component';
+import { SocialMediaComponent } from '../components/social-media.component';
 
 @NgModule({
   imports: [
@@ -14,7 +17,17 @@ import { MODULE_ROUTES_COMPONENTS, MODULE_ROUTES } from '../app-routing.module';
     RouterModule.forChild(MODULE_ROUTES),
     CarouselModule
   ],
-  declarations: [MODULE_ROUTES_COMPONENTS],
-  exports: [MODULE_ROUTES_COMPONENTS]
+  declarations: [
+    MODULE_ROUTES_COMPONENTS,
+    ListenLiveComponent,
+    SmashTeamComponent,
+    SocialMediaComponent
+  ],
+  exports: [
+    MODULE_ROUTES_COMPONENTS,
+    ListenLiveComponent,
+    SmashTeamComponent,
+    SocialMediaComponent
+  ]
 })
 export class HomeModule { }
